+++ crates/bin/streamit/src/lib.rs
@@ -1,6 +1,5 @@
 #[allow(clippy::missing_errors_doc)]
 //
-//
 pub mod configure_tracing;
 pub mod message;
 pub mod topic;
