+++ crates/bin/streamit/src/main_consumer.rs
@@ -1,13 +1,10 @@
 use anyhow::{Context, Result};
 use async_trait::async_trait;
 use bilrost::BorrowedMessage;
-use fluvio::{
-  consumer::{ConsumerConfigExtBuilder, ConsumerStream, Record as ConsumerRecord},
-  Fluvio, Offset,
-};
-use fluvio_protocol::link::ErrorCode;
+use consumer::{Consumer, FluvioConsumer};
+use fluvio::{consumer::Record as ConsumerRecord, Offset};
 use futures::StreamExt;
-use std::{pin::Pin, sync::Arc, time::Duration};
+use std::{sync::Arc, time::Duration};
 use streamitlib::{
   configure_tracing::init,
   message::{MessageKind, MessageWrapper},
@@ -19,6 +16,9 @@ use tokio::{
   time::sleep,
 };
 use tracing::{debug, error, info, trace, warn};
+
+mod consumer;
+mod mock_consumer;
 
 #[tokio::main]
 async fn main() {
@@ -89,53 +89,16 @@ impl Pinger for RealPinger {
   }
 }
 
-#[async_trait]
-trait Consumer: Send + Sync {
-  async fn consume(
-    &self,
-    topic: &str,
-    offset: Offset,
-  ) -> std::result::Result<Pin<Box<dyn ConsumerStream<Item = Result<ConsumerRecord, ErrorCode>> + Send>>, anyhow::Error>;
-}
-
-struct FluvioConsumer {}
-
-#[async_trait]
-impl Consumer for FluvioConsumer {
-  async fn consume(
-    &self,
-    topic: &str,
-    offset: Offset,
-  ) -> Result<Pin<Box<dyn ConsumerStream<Item = Result<ConsumerRecord, ErrorCode>> + Send>>, anyhow::Error> {
-    Ok(Box::pin(
-      Fluvio::connect()
-        .await?
-        .consumer_with_config(
-          ConsumerConfigExtBuilder::default()
-            .topic(topic)
-            .partition(0)
-            .offset_start(offset)
-            .build()?,
-        )
-        .await?,
-    ))
-  }
-}
-
 async fn receiver(
   tx: &Sender<ConsumerRecord>,
   pinger: &(impl Pinger + ?Sized),
   consumer: Arc<impl Consumer + ?Sized>,
 ) -> anyhow::Result<()> {
   // TODO commit offset and set consumer name https://github.com/infinyon/fluvio/blob/master/rfc/offset-management.md
-  trace!("Starting to consume");
 
   let mut stream = consumer.clone().consume(MYIO_TOPIC, Offset::beginning()).await.unwrap();
 
-  trace!("Waiting for messages");
-
   while let Some(msg) = stream.next().await {
-    trace!("Received message");
     match msg {
       Ok(msg) => tx.send(msg).await.context("Failed to send to the mpsc channel")?,
       Err(e) => error!("{e:?}"),
@@ -178,13 +141,9 @@ async fn handle_signals() -> anyhow::Result<()> {
 #[cfg(test)]
 mod tests {
   use super::*;
+  use crate::mock_consumer::tests::MockConsumer;
   use bilrost::Message;
-  use fluvio_protocol::record::{Batch, Record, RecordData};
-  use futures::{stream, Stream};
-  use std::{
-    sync::atomic::{AtomicBool, Ordering},
-    task::ready,
-  };
+  use std::sync::atomic::{AtomicBool, Ordering};
   use streamitlib::message::Birth;
   use tracing_test::traced_test;
 
@@ -200,86 +159,6 @@ mod tests {
         self.called.store(true, Ordering::SeqCst);
       }
       "pong".to_string()
-    }
-  }
-
-  // Mock Consumer that emits a single message
-  #[derive(Clone)]
-  struct MockConsumer {
-    record_values: Vec<Record>,
-  }
-
-  impl MockConsumer {
-    fn new(record_values: Vec<impl Into<RecordData>>) -> Self {
-      Self {
-        record_values: record_values
-          .into_iter()
-          .map(|record_value| Record::new(record_value.into()))
-          .collect(),
-      }
-    }
-  }
-
-  #[async_trait]
-  impl Consumer for MockConsumer {
-    async fn consume(
-      &self,
-      _topic: &str,
-      _offset: Offset,
-    ) -> Result<Pin<Box<dyn ConsumerStream<Item = Result<ConsumerRecord, ErrorCode>> + Send>>, anyhow::Error> {
-      info!("MockConsumer::consume");
-
-      let mut batch = Batch::new();
-      self
-        .record_values
-        .iter()
-        .for_each(|record| batch.add_record(record.clone()));
-
-      // we could also do this if we change &self to &mut self and pass the consumer around as
-      // Arc<Mutex<dyn Consumer> so that we have a mutable self.record_values
-      // batch.add_records(&mut self.record_values);
-
-      let stream = SinglePartitionConsumerStream {
-        // inner: stream::iter(vec![Ok(record)]),
-        inner: stream::iter(batch.into_consumer_records_iter(0).map(Ok)),
-      };
-      Ok(Box::pin(stream))
-    }
-  }
-
-  pub struct SinglePartitionConsumerStream<T> {
-    inner: T,
-  }
-
-  impl<T: Stream<Item = Result<ConsumerRecord, ErrorCode>> + Unpin> ConsumerStream for SinglePartitionConsumerStream<T> {
-    fn offset_commit(&mut self) -> std::result::Result<(), ErrorCode> {
-      todo!()
-    }
-
-    fn offset_flush(&mut self) -> futures::future::BoxFuture<'_, std::result::Result<(), ErrorCode>> {
-      todo!()
-    }
-  }
-
-  impl<T: Stream<Item = Result<ConsumerRecord, ErrorCode>> + Unpin> Stream for SinglePartitionConsumerStream<T> {
-    type Item = Result<ConsumerRecord, ErrorCode>;
-
-    fn poll_next(
-      self: std::pin::Pin<&mut Self>,
-      cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Option<Self::Item>> {
-      let self_mut = self.get_mut();
-      let pinned = std::pin::pin!(&mut self_mut.inner);
-      match ready!(pinned.poll_next(cx)) {
-        Some(Ok(last)) => {
-          trace!("LAST");
-          std::task::Poll::Ready(Some(Ok(last)))
-        }
-        other => {
-          trace!("OTHER");
-          std::task::Poll::Ready(other)
-        }
-      }
     }
   }
